mxgmn import exploration

+++ src/graph.rs
@@ -57,13 +57,13 @@ impl WaveFunction {
     }
 
     /// Leaves a random bit set to 1 and the rest to 0
-    pub fn select_random<R: Rng>(&mut self, rng: &mut R, weights: &Vec<u32>) {
+    pub fn select_random<R: Rng>(&mut self, rng: &mut R, weights: &Vec<f32>) {
         let mut weighted_rng = WeightedIndex::new(weights).unwrap();
         for i in 0..TILE_U32S {
             for j in 0..32 {
                 let index = i * 32 + j;
                 if self[i] & (1 << j) == 0 && index < weights.len() {
-                    weighted_rng.update_weights(&[(index, &0)]).unwrap();
+                    weighted_rng.update_weights(&[(index, &0.0)]).unwrap();
                 }
             }
         }
