+++ src/lib.rs
@@ -21,7 +21,7 @@ pub type UserData = Option<Box<dyn Any + Send + Sync>>;
 pub struct Peasant {
     pub graph: Graph<WaveFunction>,
     pub constraints: Arc<Vec<Vec<WaveFunction>>>,
-    pub weights: Arc<Vec<u32>>,
+    pub weights: Arc<Vec<f32>>,
     pub seed: u64,
     pub user_data: UserData,
 }
